feat: add navigation bar and login form
The navigation containing the login form, logout link and a link to the dashboard
was added and styled in accordance to the blog's theme.

The login form is only visible to visitors that are not logged in.
The logout link and the link to the private dashboard are only
visible to users that are logged in.

The functionality of the login form is yet to be programmed.