feat: secure page access and check for valid login
The homepage can now continue an existing session or create a new one,
if the corresponding cookie cannot be found.

Upon loading the page, the login status is checked. If the user has not
a valid login, the started session is destroyed. In the case of a valid
login, the user sees a custom navigation.