feat: Add functionality to the login form
The login information that the user inputs into the login form
gets sanitized to remove potentially harmful code and validated
formally, then it is checked whether the user exists in the database
and has input the matching password.

If the data matches with the data from the userbase, the user is logged
in and a session is created for them via a cookie.

If anything goes wrong during the login process, the user receives an
appropriate notification.

Once the user is logged in, they are forwarded to the dashboard.