feat: Add a logout functionality on the homepage
When the user clicks on the logout link on the homepage,
the parameters get processed and sanitized from potentially
harmful code.

The current session is then ended and the
homepage reloaded. The user is now logged out and is
required to log in again, should they wish to visit the
dashboard.