feat: Set up dashboard
The dashboard got a nagivation, a header and a footer matching
the ones on the homepage.

The navigation contains a functioning link to the homepage and a logout link
(funtionality to be added).

The page access to the dashboard was secured. Only logged in users may see
the dashboard.
The website's cookie is being read and the user redirected to the homepage if the
cookie information is missing or invalid.
Otherwise the user ID is pulled from the corresponding file.