feat: Add logout functionality to the dashboard
When the user clicks on the logout link on the dashboard,
the parameters get processed and sanitized from potentially
harmful code.

The current session is then ended and the
user is redirected to the homepage. The user is now logged out and is
required to log in again, should they wish to visit the
dashboard.