feat: Create forms in dashboard for adding categories and articles
A form was created and styled to add a new category.

A form was created and styled to add a blog article that contains
a category, title, the main text and optionally an image.
The author may choose how to align the image.

Functionality of the forms to be added...