feat: Process category form and display categories dynamically
The list of categories displayed in the select field of the article form
is now pulled from the database.

The new category that is input by the user in the category form is received,
sanitized against potentially harmful code and validated formally.
Then it is checked whether the category already exists in the database.
In that case the user is notified.
Otherwise the category is saved to the database and the category list
is pulled again from the database to be displayed in the select field
of the article form.

The user is notified whether saving the category was successful or not.