feat: Process article form and save data to db
The form data is being sanitized against potentially harmful
code and the fields are being validated.
The optional image upload is being validated.
Error notifications are thrown for arising issues.
If everything checks out, the form data is uploaded to the
database and the form is reset.
The user receives a success notification.