feat: Create popup notifications
When an error occurs upon saving a new category or
blog article or when the saving is successful, a
notification pops up for the user telling them about
the result. This way the user doesn't have to search
for this feedback on the page.

When an error occurs trying to save a new category, the entered
category that caused the error is still displayed in the field.
The form clears if saving the new category was successful.