feat: Display blog articles on the homepage
Blog articles that have been saved to the database are not
displayed dynamically on the homepage.

In addition to the blog article itself and an optional image,
author data and a timestamp are displayed as well.
The corresponding category is included.

If an image was saved along with the article, it is displayed
on the left or the right of the text, according to what was specified when submitting
the article. The text wraps around it.