perf: Reduce amount of DB operations when filtering
The code was refactored for filtering the blog articles by
category to only need one database operation instead of two
to reduce work load.

This was achieved by working with two URL parameters and
using the second one as a flag to branch the database operation
that pulls the blog article data when the page is loaded.