feat: Add an error log for admin
An error will be written into an error log for the admin to see,
if a category or blog post cannot be saved to the database.

Should the directory for the error log or the error log itself
not exist yet, then they will be created when an error occurs
and the error message will be written into the log.