feat: Add a delete function to the blog
The dashboard has received a new form to view, edit and
delete blog posts in the database.

Posts can be selected by title and deleted. The form gets
updated immediately.

Also, when a new post is created, the selection of blog posts
in the edit and delete form get updated immediately as well.