feat: Add ability to edit a blog post
When the user chooses to edit a blog post in the edit and delete
form, the form to write a new blog post is replaced by the form to
edit existing posts.

All pieces of information can be edited there, including the image. If the image is
replaced by a different one, the original image is deleted from the
server.

Should the user wish to return to writing a new blog post, a link is
present to do that.