feat: Make footer dynamic and test website
The date in the footer has been made dynamic depending
on the current year.

The website has been thoroughly tested for functionality
and any arising bugs have been dealt with.

An overlay has been created for the popup messages to
prevent the user to interact with the site despite
the popup.